requirements.txt
PyQt6==6.11.0
python-docx==1.2.0
docxtpl==0.20.2
Pillow==12.2.0